server/requirements.txt
asgiref==3.7.2
boto3==1.34.2
botocore==1.34.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.6
django-allauth==0.50.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-pandas==0.6.6
django-rest-framework==0.1.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
joblib==1.3.2
kafka-python==2.0.2
lunardate==0.2.2
mysqlclient==2.2.0
numpy==1.26.2
oauthlib==3.2.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.0
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytimekr==0.1.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.9.0
scikit-learn==1.3.2
scikit-surprise==1.1.3
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
